Moved AzBatch executor from executors/__init__ to its own file
